backend/requirements.txt
asgiref==3.5.2
astroid==2.12.12
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
autopep8==1.7.0
backports.zoneinfo==0.2.1
cffi==1.15.1
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
dill==0.3.6
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.9.1
pycparser==2.21
pylint==2.15.5
pyOpenSSL==23.1.1
pytz==2022.5
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
wrapt==1.14.1
zope.interface==6.0
